[BREAKING CHANGE] Consider content to include prefix/suffix
Providing them as separate arguments makes for a less-repetitive
signature, but also makes the class very inflexible when generated
content.

Change the signature to instead provide the full content and as
separate optional parameters the prefix and suffix.

Also change the default from empty string to null so that we
don't merge plain strings that have no prefix/suffix provided.

This becomes important when consuming code joins all strings with
a line break. Previously, consecutive strings with no prefix/suffix
would be incorrectly merged.

Also update the readme example to not use `<script>` because
GitHub's readme parser is stripping those when used inside `<pre>`
(as oppposed to when they're used inside backticks).

Change-Id: I9a95f6c2bd6a5be3fcef931da18e3357608bd296