Implement WrappedStringList
This allows a caller to collect an array of WrappedString objects, join them.
And then hand it off in a string-compatible manner to another caller which can
then still extend that list further.

Use case is MediaWiki core where we collect wrapped strings as multiple layers.
Right now they can only be merged within a single method because it needs to
expose it as a string instead of an array. Which means the caller loses
information.

This new class would allow OutputPage::getInlineHeadScripts() and buildCssLinks()
to return WrappedStringList which will act as before if treated as a string.
But for the common case where OutputPage::headElement() calls these two methods
it can continue to merge compatible objects.

Change-Id: I19e80ce9a4c1414bd8339cae1254062c3d0d6975