Fix handling of empty lists
Get rid of the fragile use of current() and next() which have an
ambiguity in that `false` can either be a genuine value, or
indicate an empty array, or an indicate that the end of the array
was reached.

Use foreach() instead with a simple $first tracker to keep
the logic as before (with a $prev separate from the loop).

Bug: T196496
Change-Id: If5853dafb221ccecb47d4ac6c30bb40950bbbfd2